Controller cleanup
