+++ lib/Controller/DocumentenController.php
@@ -1,52 +1,34 @@
 <?php
 
-namespace OCA\DsoNextcloud\Controller;
+namespace OCA\ZaakAfhandelApp\Controller;
 
 use GuzzleHttp\Client;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\TemplateResponse;
 use OCP\AppFramework\Http\JSONResponse;
-use OCP\IRequest;
 use OCP\IAppConfig;
-
-use OCA\DsoNextcloud\AppInfo\Application;
+use OCP\IRequest;
 
 /**
  * Geeft invulling aan https://vng-realisatie.github.io/gemma-zaken/standaard/zaken/
  */
-class ResultatenController extends Controller
+class DocumentenController extends Controller
 {
-    const TEST_ARRAY = [
-        "5137a1e5-b54d-43ad-abd1-4b5bff5fcd3f" => [
-            "id" => "5137a1e5-b54d-43ad-abd1-4b5bff5fcd3f",
-            "name" => "Github",
-            "summary" => "summary for one"
-        ],
-        "4c3edd34-a90d-4d2a-8894-adb5836ecde8" => [
-            "id" => "4c3edd34-a90d-4d2a-8894-adb5836ecde8",
-            "name" => "Gitlab",
-            "summary" => "summary for two"
-        ],
-        "15551d6f-44e3-43f3-a9d2-59e583c91eb0" => [
-            "id" => "15551d6f-44e3-43f3-a9d2-59e583c91eb0",
-            "name" => "Woo",
-            "summary" => "summary for two"
-        ],
-        "0a3a0ffb-dc03-4aae-b207-0ed1502e60da" => [
-            "id" => "0a3a0ffb-dc03-4aae-b207-0ed1502e60da",
-            "name" => "Decat",
-            "summary" => "summary for two"
-        ]
-    ];
 
-    public function __construct(
-		$appName,
+	/**
+	 * @var IConfig
+	 */
+	private $config;
+
+
+	public function __construct(
+		string $appName,
 		IRequest $request,
-		private readonly IAppConfig $config
-	)
-    {
-        parent::__construct($appName, $request);
-    }
+		IAppConfig $config
+	) {
+		parent::__construct($appName, $request);
+		$this->config = $config;
+	}
 
 	/**
 	 * This returns the template of the main app's page
@@ -57,7 +39,7 @@ class ResultatenController extends Controller
 	 *
 	 * @return TemplateResponse
 	 */
-	public function pages(): TemplateResponse
+	public function page(): TemplateResponse
 	{			
         return new TemplateResponse(
             //Application::APP_ID,
